Fix rm bug

+++ apothecary/formulas/assimp.sh
@@ -319,7 +319,9 @@ function copy() {
     fi
 
     # copy license files
-    rm -rf $1/license # remove any older files if exists
+    if [ -d "$1/license" ]; then
+        rm -rf $1/license
+    fi
     mkdir -p $1/license
     cp -v LICENSE $1/license/
 }
