+++ apothecary/formulas/curl/curl.sh
@@ -374,7 +374,9 @@ function copy() {
 	fi
 
 	# copy license file
-	rm -rf $1/license # remove any older files if exists
+    if [ -d "$1/license" ]; then
+        rm -rf $1/license
+    fi
 	mkdir -p $1/license
 	cp -v COPYING $1/license/
 }
